p3: Massive refactor out into classes for the connection handling
Changes: Basically put the connection related functions into its own class, and added states and state transitions for each scene. Added a snapping state for when the user drags scenes in range of connection with other scenes.
Reasons: The code was getting complicated and overly conditional in the onDrag handler function. It was not scalable with the second part of the project, which is supposed to be about sequencing scenes together.