p3: Added temporary snap state transition lever to linked state
Changes: Added a temporary global lever bool variable (createLink) setup in Main.js that listens to a click button callback on the create link button div. When this var is true, the snapped states know inside the system to transition to linked state.

Reasons: Need to move on to iterate on linked state behaviour for scenes. Will be refactored soon. Also was the simplest solution for now, as only each scene's momentFSM's states know when to respond to this callback without needing to query information from outside the system.