p3: Fixed the create link button
Changes: Made it a phaser object, and fixed the event triggering on it.
Moved the doubly linked list stuff in the enter step of LinkedState.js, as it's only needed to be setup once.
Fixed the wrong scene references to Controller.js properties (Firefox debugger is nice!). Reset referenced variables for currentlydraggedscene and neighbour after breaking away as well from Snapped State. The create button now is based on availableConnections being true instead of currentlydraggedscene and neighbour being defined.

Reasons: Fixes needed to check for creating links between scenes.