p3: Refactored out the linking scenes methods into the controller
Changes: Moved the methods in the linked state into the controller. Minor refactors in the controller for clarity.

Reasons: The methods were supposed to be called only once in the controller (was calling each method 4 times for 2 scenes active, as each called its own dragHandler and closestNeighbour from the controller in their own linked state. In short, caching the dragHandler and closestNeighbour in the controller made it inconsistent with checking each time these parameters in the linked states to do things then.