p3: Added further enter scene condition (has to be currently dragged scene)
Changes: Added condition to enter a new scene; that scene has to be the currently dragged one. 

Reasons: Simplifies handling scene traversal from one to the other by just getting the closest neighbour to that scene which was entered by the player. Hopefully prevents logical errors for handling scenes and knowing which one is active as well. Lastly, makes the interaction with the dragged scene more meaningful.