p3: Added b/w tilemap to World.js, assets, dom
Changes: Added assets. This is a WIP commit before changing scenes into game objects, so at this point things are not configured yet.

Reasons: The project became a top down game with interesting mechanics. Playing around with the DOM led to a decision to make it a game-based psychological questionnaire experiment. 

Previously, using scenes that were draggable were fine with the whole connect moments idea, until it became an embodied experience. The data visualization was for a d3 thing which does not fit the current scope and project goals (also beyond my skills, as it needed advanced data structures and algorithms).